Implemented automatic algorithm for schedule issue #288

+++ API/Application/DTO/SchedulingDTO.cs
@@ -5,13 +5,13 @@ public class SchedulingDTO
 {
     public List<SchedulingEntryDTO> Entries { get; set; }
 
-    public int TotalDelay { get; set; }
+    public double TotalDelay { get; set; }
 
     public double ExecutionTime { get; set; }
 
     public List<string> Messages { get; set; }
 
-    public SchedulingDTO(List<SchedulingEntryDTO> entries, int totalDelay, double executionTime, List<string>? messages = null)
+    public SchedulingDTO(List<SchedulingEntryDTO> entries, double totalDelay, double executionTime, List<string>? messages = null)
     {
         Entries = entries;
         TotalDelay = totalDelay;
